+++ bindings/hyperdrivepy/Cargo.toml
@@ -26,4 +26,4 @@ fixed-point-macros = { path = "../../crates/fixed-point-macros" }
 hyperdrive-math = { path = "../../crates/hyperdrive-math" }
 hyperdrive-wrappers = { path = "../../crates/hyperdrive-wrappers" }
 
-pyo3 = { version = "0.19.0", features = ["abi3-py37"] }
+pyo3 = { version = "0.19.0", features = ["abi3-py37", "multiple-pymethods"] }
